orderModel++

+++ backend/order-service/src/main/java/com/buyone/orderservice/model/OrderItem.java
@@ -2,6 +2,8 @@ package com.buyone.orderservice.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -9,8 +11,8 @@ import lombok.*;
 public class OrderItem {
     private String productId;
     private String productName;
-    private String sellerId;
-    private double price;
+    private String sellerId;    // For seller dashboards/revenue "SELLER" from user DB
+    private BigDecimal price;
     private int quantity;
-    private String imageUrl;
+    private String imageUrl;    // First image from Product.images
 }
